requirements.txt
click==6.7
Pillow==4.0.0
